+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/classification/BernoulliNaiveBayesTest.java
@@ -46,7 +46,7 @@ public class BernoulliNaiveBayesTest extends AbstractTest {
     public void testPredict() {
         logger.info("testPredict");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         
         Dataframe[] data = Datasets.carsNumeric(configuration);
@@ -91,7 +91,7 @@ public class BernoulliNaiveBayesTest extends AbstractTest {
     public void testKFoldCrossValidation() {
         logger.info("testKFoldCrossValidation");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         int k = 5;