+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/clustering/KmeansTest.java
@@ -44,7 +44,7 @@ public class KmeansTest extends AbstractTest {
     public void testPredict() {
         logger.info("testPredict");
 
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         
         Dataframe[] data = Datasets.heartDiseaseClusters(configuration);
@@ -116,7 +116,7 @@ public class KmeansTest extends AbstractTest {
     public void testKFoldCrossValidation() {
         logger.info("testKFoldCrossValidation");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         int k = 5;