+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/clustering/MultinomialDPMMTest.java
@@ -43,7 +43,7 @@ public class MultinomialDPMMTest extends AbstractTest {
     public void testPredict() {
         logger.info("testPredict");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         Dataframe[] data = Datasets.multinomialClusters(configuration);
         
@@ -88,7 +88,7 @@ public class MultinomialDPMMTest extends AbstractTest {
     public void testKFoldCrossValidation() {
         logger.info("testKFoldCrossValidation");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         int k = 5;