+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/featureselection/TFIDFTest.java
@@ -42,7 +42,7 @@ public class TFIDFTest extends AbstractTest {
     public void testSelectFeatures() {
         logger.info("selectFeatures");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         Dataframe[] data = Datasets.featureSelectorTFIDF(configuration);