+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/recommendation/CollaborativeFilteringTest.java
@@ -45,7 +45,7 @@ public class CollaborativeFilteringTest extends AbstractTest {
     public void testPredict() {
         logger.info("testPredict");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         Dataframe[] data = Datasets.recommenderSystemFood(configuration);