+++ datumbox-framework-core/src/test/java/com/datumbox/framework/core/machinelearning/regression/StepwiseRegressionTest.java
@@ -42,7 +42,7 @@ public class StepwiseRegressionTest extends AbstractTest {
     public void testPredict() {
         logger.info("testPredict");
         
-        Configuration configuration = Configuration.getConfiguration();
+        Configuration configuration = getConfiguration();
         
         Dataframe[] data = Datasets.regressionNumeric(configuration);